quantity datatype changed to decimal, 3 digit after point

+++ database/migrations/2023_10_19_074428_create_order_products_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreign('order_id')->references('id')->on('orders');
             $table->unsignedBigInteger('product_id');
             $table->foreign('product_id')->references('id')->on('products');
-            $table->integer('quantity');
+            $table->decimal('quantity', 15, 3);
             $table->decimal('price', 15, 2);
             $table->decimal('total_price', 15, 2);
             $table->integer('created_by')->unsigned()->nullable();
